requirements.txt
asgiref==3.5.0
Django==4.0.2
fpdf==1.7.2
lxml==4.7.1
Pillow==9.0.1
pytesseract==0.3.8
python-docx==0.8.11
sqlparse==0.4.2
tzdata==2021.5
